+++ src/model/hazard.ts
