+++ src/route/trail.ts
@@ -1,8 +1,8 @@
 import { Context, Hono } from "hono";
-import { TrailList } from "../services/trails_service.ts";
+import { TrailList } from "../service/trails_service.ts";
 import logger from "../logger.ts";
 import * as decorators from "../decorators.ts";
-import trailsService from "../services/trails_service.ts";
+import trailsService from "../service/trails_service.ts";
 
 const app = new Hono()
   .get("/list", (ctx: Context) => {
